Add menu option on post-process mode to process/reprocess loaded data. Required some other refactoring and additions to make it work cleanly and intuitively.
- Move calculation of curtain display to when user clicks CurtainPlot so it will be recalculated if user has reprocessed data.
- Update test for if data is processed to be able to test any scan requested, not just first scan.
- Implement deep copy of structures BestWtdRCSet and ATP that hold processed data so each scan has it’s own struct and doesn’t step on each other. This did not affect realtime processing, just save and restore from JSON file.

Add ability for user to type a scan number into the Scan Index field and click Go to jump to that scan.  Required tweaking the layout of the main display window a bit.

Fix quit button so works when processor window is open.

Change return values of a couple functions to be more useful/informative.

Add/clarify a few comments and displayed text.

Add/modify unit tests for changes.